latest updates

+++ Server/Controllers/Assignement.js
@@ -6,12 +6,12 @@ const ls=require('local-storage')
 
 
 const creatAssignement= async(req,res)=>{
-   const arr={users:["63bed9952daa0a2cd91bae02","63c0d9d3aac7a477db8a539c"],cours:"63bf2f32831337ee4b6e8f97"}
-for(let i=0;i<arr.users.length;i++){
-        let body={user:arr.users[i],cours:arr.cours,statu:arr.status} 
+    const {users,cours}=req.body
+for(let i=0;i<users.length;i++){
+        let body={user:users[i],cours:cours} 
        const oldAssignementrows= await Assignement.findOne({user:body.user}).count('_id')
        const oldAssignement= await Assignement.findOne({user:body.user}).populate({path:'user',model:User}).populate({path:'cours',model:Cours})
-       const newAssignementCours= await Cours.findOne({_id:arr.cours})
+       const newAssignementCours= await Cours.findOne({_id:cours})
           if(oldAssignementrows<2) { 
                if(!oldAssignement) await Assignement.create({...body}) 
                  if(oldAssignement)
@@ -19,13 +19,13 @@ for(let i=0;i<arr.users.length;i++){
                      if((oldAssignement.cours.start.getTime()<newAssignementCours.start.getTime() && oldAssignement.cours.end.getTime()<newAssignementCours.start.getTime()) || (oldAssignement.cours.start.getTime()>newAssignementCours.end.getTime() && oldAssignement.cours.end.getTime()>newAssignementCours.end.getTime())){
                          await Assignement.create({...body})
                         }else{
-                          RemoveAssignedRowsByUser(arr.users,arr.cours)
-                          return res.json({errmsg:`${oldAssignement.user.name} a dejas une formation dans cet intervale de temp`, assigned_to:done}) 
+                          RemoveAssignedRowsByUser(users,cours)
+                          return res.json({errmsg:`${oldAssignement.user.name} a dejas une formation dans cet intervale de temp`}) 
                          }
                  }}
           else{
-        RemoveAssignedRowsByUser(arr.users,arr.cours)
-        return res.json({errmsg: `${oldAssignement.user.name} a dejas deux formations a valider`,assigned_to:done})
+        RemoveAssignedRowsByUser(users,cours)
+        return res.json({errmsg: `${oldAssignement.user.name} a dejas deux formations a valider`})
              }
     }
     res.json({msg:'assigne'})
@@ -41,24 +41,27 @@ const ValidityChecker= ()=>async (req,res)=>{
     const assignements= await Assignement.find()
     assignements.forEach(async (a)=>{
         const assignement= await Assignement.findOne({_id:a._id}).populate({path:'cours',model:Cours})
+           if(assignement) {
             if(assignement.cours.end.getTime()<Date.now()){
-                const saved= await Historique.create({user:assignement.user._id,cours:assignement.cours._id})
-                const removed= await Assignement.findOneAndRemove({user:assignement.user._id,cours:assignement.cours._id})
+               await Historique.create({user:assignement.user._id,cours:assignement.cours._id})
+               await Assignement.findOneAndRemove({user:assignement.user._id,cours:assignement.cours._id})
             }
+
+        }
       })
       res.json({msg:'saved'})
 }
 
 const getAssignementes= async(req,res)=>{
     const {body}=req
-    const Assignementes= await Assignement.find()
+    const Assignementes= await Assignement.find().populate({path:'user',model:User}).populate({path:'cours',model:Cours})
     if(!Assignementes) throw Error('Assignement not found')
     res.json(Assignementes)
 }
 const getAssignement= async(req,res)=>{
-    const {body}=req
-    const oneAssignement= await Assignement.findOne({_id:body._id})
-    if(!oneAssignement) throw Error('admin not added')
+    const {id}=req.body
+    const oneAssignement= await Assignement.findOne({user:id}).populate({path:'cours',model:Cours})
+    if(!oneAssignement) throw Error('assignement not found')
     res.json(oneAssignement)
 }
 const UpdateAssignement= async(req,res)=>{
